iniciei alterações de organização de pastas para conseguir uma arquitetura exagonal, ainda falta configurar a application.module e domain.modeule corretamente para testar o projeto
