+++ src/app.module.ts
@@ -1,13 +1,10 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { ClientsModule } from './clients/clients.module';
-import { AccountsModule } from './accounts/accounts.module';
-import { ManagersModule } from './managers/managers.module';
+import { ApplicationModule } from './application/application.module';
+import { DomainModule } from './domain/domain.module';
 
 @Module({
-  imports: [ClientsModule, AccountsModule, ManagersModule],
-  controllers: [AppController],
-  providers: [AppService],
+  imports: [ApplicationModule, DomainModule],
+  controllers: [],
+  providers: [],
 })
 export class AppModule {}
