+++ src/application/controllers/managers.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
-import { ManagersService } from './managers.service';
-import { CreateManagerDto } from './dto/create-manager.dto';
-import { UpdateManagerDto } from './dto/update-manager.dto';
-import { Manager } from './entities/manager.entity';
+import { ManagersService } from '../../domain/services/managers.service';
+import { CreateManagerDto } from '../dtos/manager/create-manager.dto';
+import { UpdateManagerDto } from '../dtos/manager/update-manager.dto';
+import { Manager } from '../../domain/entities/manager.entity';
 
 @Controller('managers')
 export class ManagersController {
