+++ src/application/dtos/account/create-account.dto.ts
@@ -1,4 +1,4 @@
-import { Client } from "src/clients/entities/client.entity";
+import { Client } from "src/domain/entities/client.entity";
 
 export class CreateAccountDto {
     private _client: Client;
