+++ src/application/dtos/account/update-account.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateAccountDto } from './create-account.dto';
+import { CreateAccountDto } from '../../dtos/account/create-account.dto';
 
 export class UpdateAccountDto extends PartialType(CreateAccountDto) {}
