+++ src/application/dtos/manager/update-manager.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateManagerDto } from './create-manager.dto';
+import { CreateManagerDto } from '../../dtos/manager/create-manager.dto';
 
 export class UpdateManagerDto extends PartialType(CreateManagerDto) {}
