+++ src/domain/services/managers.service.ts
@@ -1,32 +1,47 @@
 import { Injectable } from '@nestjs/common';
-import { Client } from './entities/client.entity';
-import { Account } from 'src/accounts/entities/account.entity';
+import { UpdateManagerDto } from '../../application/dtos/manager/update-manager.dto';
+import { Manager } from '../entities/manager.entity';
+import { Client } from '../entities/client.entity';
+import { Account } from '../entities/account.entity';
 
 @Injectable()
-export class ClientsService {
+export class ManagersService {
+    private managers: Manager[] = [];
     private clients: Client[] = [];
 
     //CRUD
-    create(client: Client): Client {
+    createNewManager(manager: Manager): Manager {
+        this.managers.push(manager);
+        return manager;
+    }
+
+    findAllManager(): Manager[] {
+        return [...this.managers];
+    }
+
+    findManagerById(id: string): Manager {
+        return this.managers.find((manager) => manager.id === id);
+    }
+
+    update(id: string, updateManagerDto: UpdateManagerDto) {
+        return `This action updates a #${id} manager`;
+    }
+
+    removeManager(id: string): Manager[] {
+        return this.managers.filter((manager) => manager.id !== id);
+    }
+
+
+    //MÉTODOS DO GERENTE
+    addClientToManager(client: Client): void {
         this.clients.push(client);
+    }
+
+    removedClientToManager(id: string, client:Client): Client {
+        this.clients.filter((client) => client.id !== id);
         return client;
     }
 
-    findAll(): Client[] {
-        return [...this.clients];
-    }
-
-    findOne(id: string): Client {
-        return this.clients.find((client) => client.id === id);
-    }
-  
-    remove(id: string): Client[] {
-        return this.clients = this.clients.filter((client) => client.id !== id);
-    }
-
-
-    //MÉTODOS DO CLIENTE
-    
     openAccount(account: Account, client: Client): string {
         if (client.income >= 500) {
             account.typeAccount = 'CORRENTE';
@@ -57,5 +72,4 @@ export class ClientsService {
             return account.typeAccount = 'CORRENTE';
         }
     }
-
 }
